workshop-3/aula1/projects/backend/requirements.txt

fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.5.0
stellar-sdk==13.0.0
coloredlogs==15.0.1
python-multipart==0.0.20
pydantic-settings==2.7.0
python-dotenv==1.0.1